Moved API into webui.py, added new requirements, added api flag to shared.py
