--- api_requirements.txt
@@ -1,38 +1,0 @@
-transformers~=4.19.2
-diffusers~=0.2.4
-basicsr~=1.3.5
-gfpgan
-gradio~=3.3.1
-numpy~=1.23.3
-Pillow~=9.2.0
-realesrgan
-torch~=1.12.1+cu113
-transformers
-omegaconf~=2.1.1
-pytorch_lightning
-diffusers
-invisible-watermark
-scikit-image>=0.19
-fonts~=0.0.3
-font-roboto
-git+https://github.com/crowsonkb/k-diffusion.git
-git+https://github.com/TencentARC/GFPGAN.git@8d2447a2d918f8eba5a4a01463fd48e45126a379
-timm==0.4.12
-fairscale==0.4.4
-piexif~=1.1.3
-
-fastapi~=0.85.0
-tqdm~=4.64.1
-einops~=0.4.1
-opencv-python~=4.6.0.66
-torchvision~=0.13.1
-PyYAML~=6.0
-lmdb~=1.3.0
-requests~=2.25.1
-lpips~=0.1.4
-setuptools~=59.8.0
-gdown~=4.5.1
-scipy~=1.9.1
-clip~=1.0
-kornia~=0.6.7
-packaging~=21.3
