+++ requirements.txt
@@ -20,3 +20,7 @@ git+https://github.com/TencentARC/GFPGAN.git@8d2447a2d918f8eba5a4a01463fd48e4512
 timm==0.4.12
 fairscale==0.4.4
 piexif
+
+uvicorn
+fastapi
+pydantic
