+++ requirements_versions.txt
@@ -16,3 +16,7 @@ font-roboto
 timm==0.4.12
 fairscale==0.4.4
 piexif==1.1.3
+
+uvicorn==0.18.3
+fastapi==0.85.0
+pydantic==1.10.2
